Refactor to dump internal logs
